refactor(http): split handler.rs into state/routes/notifications/dispatch
Break the 688-line crates/dcc-mcp-http/src/handler.rs monolith into a

thin facade backed by four focused siblings:

- handler_state.rs: AppState + cancellation/elicitation TTL constants

- handler_routes.rs: handle_post / handle_get / handle_delete

- handler_notifications.rs: notification + elicitation-response routing

- handler_dispatch.rs: JSON-RPC method router + initialize + tools/list

Behaviour unchanged; handler.rs re-exports every symbol so existing

call sites (both inside the crate and in src/handlers/*) keep using

crate::handler::* unchanged. Updates docs/guide/architecture.md with

the new maintainer layout.
