+++ crates/dcc-mcp-http/src/handler_routes.rs
@@ -1,0 +1,210 @@
+//! Axum request handlers for the three MCP Streamable HTTP verbs.
+//!
+//! - `POST /mcp`   — client sends JSON-RPC messages; response is JSON or SSE
+//! - `GET  /mcp`   — client opens a long-lived SSE stream for server-push events
+//! - `DELETE /mcp` — client closes its session
+
+use axum::{
+    body::Body,
+    extract::State,
+    http::{HeaderMap, StatusCode, header},
+    response::sse::Event,
+    response::{IntoResponse, Response, Sse},
+};
+use futures::stream;
+use serde_json::Value;
+use tokio::sync::broadcast;
+use tokio_stream::StreamExt;
+use tokio_stream::wrappers::BroadcastStream;
+
+use super::dispatch::dispatch_request;
+use super::notifications::{handle_notification, handle_response_message};
+use super::state::AppState;
+use crate::handlers::{json_error_response, json_has_id, parse_body, parse_raw_values};
+use crate::protocol::{
+    self, JsonRpcBatch, JsonRpcMessage, JsonRpcResponse, MCP_SESSION_HEADER, format_sse_event,
+};
+
+/// Handle `POST /mcp`: accept JSON-RPC message(s) and return response.
+pub async fn handle_post(
+    State(state): State<AppState>,
+    headers: HeaderMap,
+    body: String,
+) -> Response {
+    let session_id = headers
+        .get(MCP_SESSION_HEADER)
+        .and_then(|v| v.to_str().ok())
+        .map(str::to_owned);
+
+    // Parse body — keep raw Value array for id-presence detection
+    let raw_values: Vec<Value> = match parse_raw_values(&body) {
+        Ok(v) => v,
+        Err(e) => {
+            return json_error_response(
+                StatusCode::BAD_REQUEST,
+                None,
+                protocol::error_codes::PARSE_ERROR,
+                format!("Parse error: {e}"),
+            );
+        }
+    };
+
+    let messages: JsonRpcBatch = match parse_body(&body) {
+        Ok(m) => m,
+        Err(e) => {
+            return json_error_response(
+                StatusCode::BAD_REQUEST,
+                None,
+                protocol::error_codes::PARSE_ERROR,
+                format!("Parse error: {e}"),
+            );
+        }
+    };
+
+    // A message is a "request" (needs a response) iff it has an explicit "id" field.
+    let has_requests = raw_values.iter().any(json_has_id);
+
+    // Always process notifications (fire-and-forget — no id) so that
+    // `notifications/cancelled` can abort in-flight tool calls.
+    for msg in &messages {
+        if let JsonRpcMessage::Notification(notif) = msg {
+            handle_notification(&state, &notif.method, notif.params.as_ref()).await;
+        }
+    }
+    // Client responses to server-initiated elicitation requests arrive as
+    // JSON-RPC responses. Correlate and wake the waiting oneshot channel.
+    for msg in &messages {
+        if let JsonRpcMessage::Response(resp) = msg {
+            handle_response_message(&state, resp);
+        }
+    }
+
+    if !has_requests {
+        // Only notifications/responses — accept and return 202
+        return StatusCode::ACCEPTED.into_response();
+    }
+
+    // Process requests and build responses
+    let mut responses: Vec<JsonRpcResponse> = Vec::new();
+    let mut use_sse = false;
+
+    // Check if client accepts SSE
+    if let Some(accept) = headers.get(header::ACCEPT) {
+        if accept.to_str().unwrap_or("").contains("text/event-stream") {
+            use_sse = true;
+        }
+    }
+
+    for msg in &messages {
+        if let JsonRpcMessage::Request(req) = msg {
+            match dispatch_request(&state, req, session_id.as_deref()).await {
+                Ok(resp) => responses.push(resp),
+                Err(e) => {
+                    responses.push(JsonRpcResponse::internal_error(
+                        req.id.clone(),
+                        e.to_string(),
+                    ));
+                }
+            }
+        }
+    }
+
+    if use_sse && session_id.is_some() {
+        // Return as SSE stream (allows server push alongside response)
+        let events: Vec<String> = responses
+            .iter()
+            .map(|r| format_sse_event(r, None))
+            .collect();
+
+        let stream = stream::iter(events).map(Ok::<_, std::convert::Infallible>);
+
+        let body = Body::from_stream(stream);
+        Response::builder()
+            .status(StatusCode::OK)
+            .header(header::CONTENT_TYPE, "text/event-stream")
+            .header("Cache-Control", "no-cache")
+            .header("X-Accel-Buffering", "no")
+            .body(body)
+            .unwrap_or_else(|_| StatusCode::INTERNAL_SERVER_ERROR.into_response())
+    } else {
+        // Return as JSON
+        let body = if responses.len() == 1 {
+            serde_json::to_string(&responses[0]).unwrap_or_default()
+        } else {
+            serde_json::to_string(&responses).unwrap_or_default()
+        };
+        Response::builder()
+            .status(StatusCode::OK)
+            .header(header::CONTENT_TYPE, "application/json")
+            .body(Body::from(body))
+            .unwrap_or_else(|_| StatusCode::INTERNAL_SERVER_ERROR.into_response())
+    }
+}
+
+/// Handle `GET /mcp`: open SSE stream for server-push events.
+pub async fn handle_get(State(state): State<AppState>, headers: HeaderMap) -> Response {
+    // Validate Accept header
+    let accept = headers
+        .get(header::ACCEPT)
+        .and_then(|v| v.to_str().ok())
+        .unwrap_or("");
+    if !accept.contains("text/event-stream") {
+        return StatusCode::METHOD_NOT_ALLOWED.into_response();
+    }
+
+    let session_id = headers
+        .get(MCP_SESSION_HEADER)
+        .and_then(|v| v.to_str().ok())
+        .map(str::to_owned);
+
+    let rx: broadcast::Receiver<String> = if let Some(id) = &session_id {
+        match state.sessions.subscribe(id) {
+            Some(rx) => rx,
+            None => {
+                return json_error_response(
+                    StatusCode::NOT_FOUND,
+                    None,
+                    -32600,
+                    "Session not found",
+                );
+            }
+        }
+    } else {
+        // No session — create an ephemeral one
+        let id = state.sessions.create();
+        state
+            .sessions
+            .subscribe(&id)
+            .expect("subscribe on a freshly created session cannot fail")
+    };
+
+    let sse_stream = BroadcastStream::new(rx)
+        .filter_map(|res| res.ok())
+        .map(|data| {
+            // Each item is already a formatted SSE event string
+            // Parse it back to send as axum SSE Event
+            Ok::<_, std::convert::Infallible>(Event::default().data(data))
+        });
+
+    Sse::new(sse_stream)
+        .keep_alive(axum::response::sse::KeepAlive::new())
+        .into_response()
+}
+
+/// Handle `DELETE /mcp`: terminate a session.
+pub async fn handle_delete(State(state): State<AppState>, headers: HeaderMap) -> StatusCode {
+    let session_id = headers
+        .get(MCP_SESSION_HEADER)
+        .and_then(|v| v.to_str().ok());
+
+    match session_id {
+        Some(id) if state.sessions.remove(id) => {
+            if state.enable_resources {
+                state.resources.drop_session(id);
+            }
+            StatusCode::NO_CONTENT
+        }
+        Some(_) => StatusCode::NOT_FOUND,
+        None => StatusCode::BAD_REQUEST,
+    }
+}
