+++ docs/guide/architecture.md
@@ -295,6 +295,7 @@ dcc-mcp-server ← dcc-mcp-http
 - `src/resources.rs` keeps the `ResourceRegistry` (producer wiring, subscription state, `notify_updated` fan-out); the `ResourceProducer` trait + content/error types live in `resources_types.rs`, the built-in producers (`scene://`, `capture://`, `audit://`, `artefact://`) live in `resources_producers.rs`, and unit tests live in `resources_tests.rs`.
 - `src/prompts.rs` keeps the `PromptRegistry` (lazy cache, skill-set invalidation, `list` / `get` surface); the YAML spec types + `PromptError` live in `prompts_spec.rs`, the `{{name}}` template engine lives in `prompts_template.rs`, the sibling-file / glob loader plus the workflow-derived prompt generator live in `prompts_loader.rs`, and unit tests live in `prompts_tests.rs`.
 - `src/protocol.rs` is a thin facade keeping protocol-version negotiation + session/method constants; every MCP message type is split by primitive: JSON-RPC envelope + error codes in `protocol_jsonrpc.rs`, lifecycle (`initialize` / capabilities / roots / logging / elicitation) in `protocol_lifecycle.rs`, tools (`tools/list` / `tools/call` / annotations / content) in `protocol_tools.rs`, resources (`resources/list|read|subscribe`) in `protocol_resources.rs`, prompts (`prompts/list|get`) in `protocol_prompts.rs`, and SSE formatter + cursor helpers in `protocol_sse.rs`.
+- `src/handler.rs` is a thin facade over the top-level axum handlers: shared `AppState` + cancellation/elicitation TTL constants live in `handler_state.rs`, the three HTTP verbs (`POST` / `GET` / `DELETE /mcp`) live in `handler_routes.rs`, notification and response-message routing (`notifications/cancelled`, `roots/list_changed`, elicitation correlation) live in `handler_notifications.rs`, and the JSON-RPC method router plus `initialize` / `tools/list` implementations live in `handler_dispatch.rs`. Per-method request handlers (`tools/call`, `resources/*`, `prompts/*`, `elicitation/create`, …) continue to live under `src/handlers/` and are re-exported through the facade so existing call sites keep using `crate::handler::*`.
 
 ### dcc-mcp-server
